backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-socketio==5.10.0
redis==5.0.1
celery==5.3.4
stripe==7.8.0
sendgrid==6.11.0
geopy==2.4.1
pytest==7.4.3
pytest-asyncio==0.21.1